WIP: round map

+++ src/main/java/ru/bulldog/justmap/util/ImageUtil.java
@@ -106,7 +106,7 @@ public class ImageUtil {
 			int x = 0;
 			while(x < width) {
 				if (imgX >= imgW) imgX = 0;
-				int len = new Line(centerX, centerY, x, y).lenght();
+				int len = (int) new Line(centerX, centerY, x, y).lenght();
 				if (len <= rOut && len >= rIn) {							
 					int pixel = texture.getPixelColor(imgX, imgY);
 					roundSkin.setPixelColor(x, y, pixel);
